REST API: updates the post-counts endpoint to include mine/all counts
This patch was developed by @retrofocs to retrieve post-counts across statuses for the current user of a site in addition to the prior existing counts across all authors. This change is in support of the project to enable dynamic filters for posts that is tracked in github here: https://github.com/Automattic/calypso-pre-oss/pull/6588. Further discussion about this commit took place here: https://[private link].

Merges r119896-wpcom.