REST API: Add autosave link to post meta.
In 'edit' context, if an autosave exists for a post that is more recent than the post's last modification, we'll link to it in meta->links.

Note that this autosave could belong to any user, which is in line with wp-admin's behavior. We may need to limit it to just the requesting user if this causes permissions headaches.

Merges r123586-wpcom.